requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.1
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
kombu==5.0.2
prompt-toolkit==3.0.8
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
